requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and RAG
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.5

# Vector Store
pinecone-client==3.0.2
chromadb==0.4.22

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
unstructured==0.12.0
markdown==3.5.2

# Embeddings and LLM
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.42.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
redis==5.0.1

# Monitoring and Logging
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
